requirements.txt
flake8
pyserial
